% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_day_summary.R
\name{get_day_summary}
\alias{get_day_summary}
\title{Get day summary as a list}
\usage{
get_day_summary(token, plant_id, date)
}
\arguments{
\item{token}{Authorisation token.}

\item{plant_id}{Plant id.}

\item{date}{Date in format YYYY-MM-DD}
}
\value{
List of power flow (and state of charge) values for the given day.
}
\description{
Get day summary as a list
}
\examples{
\dontrun{
get_day_summary(token, 1, "2024-07-30")
}
}
