% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plants.R
\name{get_plants}
\alias{get_plants}
\title{Get details of power plants associated with a Sunsynk account}
\usage{
get_plants(token)
}
\arguments{
\item{token}{Authorisation token.}
}
\value{
List.
}
\description{
Get details of power plants associated with a Sunsynk account
}
