% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get a token for authenticating to the Sunsynk API}
\usage{
get_token()
}
\value{
List
}
\description{
Get a token for authenticating to the Sunsynk API
}
\examples{
get_token()
}
