% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_day_summary_table.R
\name{plot.sunsynkr_day_summary_table}
\alias{plot.sunsynkr_day_summary_table}
\title{Plots the sunsynkr daily summary table}
\usage{
\method{plot}{sunsynkr_day_summary_table}(x, ...)
}
\arguments{
\item{x}{susynkr daily summary table.}

\item{...}{Required for S3 generics.}
}
\value{
ggplot2 plot
}
\description{
Plots the sunsynkr daily summary table
}
\examples{
\dontrun{
plot(sunsynkr_day_summary_table)
}
}
