% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plants.R
\name{print.sunsynkr_plants}
\alias{print.sunsynkr_plants}
\title{Print Sunsynk API plants}
\usage{
\method{print}{sunsynkr_plants}(x, ...)
}
\arguments{
\item{x}{\code{sunsynkr_plants} object.}

\item{...}{Required for S3 generics.}
}
\value{
A tibble of summary details for the \code{sunsynkr_plants} object.
}
\description{
Print Sunsynk API plants
}
\examples{
\dontrun{
print(sunsynkr_plants)
}
}
