% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_token.R
\name{print.sunsynkr_token}
\alias{print.sunsynkr_token}
\title{Print details of Sunsynkr token}
\usage{
\method{print}{sunsynkr_token}(x)
}
\arguments{
\item{sunsynkr}{token}
}
\description{
Print details of Sunsynkr token
}
