% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plants.R
\name{print_plants}
\alias{print_plants}
\title{Print Sunsynk API plants}
\usage{
print_plants(x)
}
\arguments{
\item{x}{Sunsynkr plants}
}
\value{
A tibble of summary details for sunsynkr plants
}
\description{
Print Sunsynk API plants
}
