% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_summary_info_to_col.R
\name{day_summary_info_to_col}
\alias{day_summary_info_to_col}
\title{Converts a daily summary 'infos' item to a dataframe}
\usage{
day_summary_info_to_col(x, date)
}
\arguments{
\item{x}{Daily summary}

\item{date}{Date to be pre-pended to the time to form a date/time field.}
}
\value{
data.frame
}
\description{
Converts a daily summary 'infos' item to a dataframe
}
