% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_day_summary_table.R
\name{get_day_summary_table}
\alias{get_day_summary_table}
\title{Get a daily summary for a plant as a data.frame}
\usage{
get_day_summary_table(token, plant_id, date)
}
\arguments{
\item{token}{Authorisation token}

\item{plant_id}{plant id}

\item{date}{Date for the summary (YYYY-MM-DD).}
}
\value{
A data.frame
}
\description{
Get a daily summary for a plant as a data.frame
}
\examples{
\dontrun{
get_day_summary_table(token, 1, "2024-07-30")
}
}
