% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_token.R
\name{print.sunsynkr_token}
\alias{print.sunsynkr_token}
\title{Print details of Sunsynkr token}
\usage{
\method{print}{sunsynkr_token}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sunsynkr_token}.}

\item{...}{Required for S3 generics.}
}
\value{
Prints details of the token to the console.
}
\description{
Print details of Sunsynkr token
}
\examples{
\dontrun{
print(token)
}
}
